@acozzette 2022-09-27
@zzachimed Thank you for the pull request, but unfortunately we are not in a good position to accept this kind of change right now. Probably the biggest problem is that we are not internally building with the necessary compiler flags to make these instructions available, so if we accepted this change then we wouldn't currently be able to test it. We have also run into problems in the past where an optimization looks great in a microbenchmark but doesn't work out in real binaries in production. This kind of optimization seems promising, but I think we will need our internal performance experts to drive this effort to make sure that we can support it properly.