@jorgbrown 2022-10-12
Thanks again for the pull request.  I really appreciate that you took the time to run benchmarks, and include them along with your code.

For what it's worth, if we were to go this route, we would want these changes at a minimum:

1) You read the next 16 bytes with _mm_lddqu_si128.  However, our input stream prefetches only 16 bytes, and that's to be sure we can read 5 bytes of tag plus 10 bytes of varint.  At the time your code runs, we may have already retrieved 5 bytes of tag information, so you may only have 11 bytes available to read... therefore you can't read 16.  Thankfully, at the time you use _mm_lddqu_si128, you've already read two bytes and incorporated them into res... so you only really need to read 8 bytes anyway.

2) _mm_mask_cmp_epi8_mask is an AVX-512 instruction.  That's a deal-breaker right there, because of the awful way that Intel can slow down the clock rate of every core on a chip just because one of them started using AVX-512.  (See https://blog.cloudflare.com/on-the-dangers-of-intels-frequency-scaling/ )  Something like _mm_movemask_epi8 , an SSE2 instruction, would be much better.

3) You're using pext.  While most x86 chips these days support it, some important ones - Zen1 and Zen2 - support it with microcode that is so slow that it's usually better not to use pext.  On a Zen it's at least twice as fast to use this for 64-bit:

// Note: Assumes the high bits of each byte of `in` are zero.
int64_t pext_7F7F7F7F7F7F7F7F(uint64_t in) {
  in += (in & 0x00FF00FF00FF00FF);
  in += (in & 0x0000FFFF0000FFFF) * 3;
  return ((0xFFFFFFFF & in) >> 3) + (in >> 32 << 25);
}

( For your perusal, I have various other algorithms for pext at https://godbolt.org/z/xvn8W4en3 )

= - = - =

Now, regarding your benchmarks, here's the really bad news...

I believe they are repeatedly testing specific widths of incoming integers.  That's very good for any branches which might be in your code, since the CPU will be able to predict them.  It's not so good for real-world parsing, however.  In the real-world, two things ruin these optimizations and others like them:

a) Most integers are single-byte encodings.  Those that are not? They are typically double-byte encodings.
b) In situations where you're dealing with larger encodings, their length is random.

You'll want to add a benchmark whose data reflects the reality of (a) and (b).  Sadly, once you've done so, you'll find it to be incredibly difficult to beat a simple loop.  In particular, if your parser is inline, then its decode will be repeated throughout the binary.  A larger faster decoder can dramatically increase the overall size of an application, and the resulting instruction cache misses can overwhelm the time savings of the faster decoder.

And that's what's happened time and time again: people with great optimization experience have come up with some great code to speed up decoding, only to find that the results of a large real-world program aren't much affected, and sometimes are actually slower.  :-(