@zzachimed 2022-11-19
@jorgbrown 👍 
  verry thanks and appreciate for your carefuly comments . sorry for latter response for current busy work. 
1 for your concern about avx512 issue 
    we disucssed with intel senior principal architect  Moga, Adrian C [adrian.c.moga@intel.com]. This is his comment about"
Intel can slow down the clock rate of every core on a chip just because one of them started using AVX-512"
"I don’t this this is true. 
First, the article is old and based on a SKX 12C SKU.

Second, the Cloudflare blog writer is confused by wikichip reference, which refers to N cores using the same license.
Nothing in there says that one core doing AVX-512 brings down the other cores to base frequency if the run SSE.

Also, if I understand his experiment correctly, he is running a fully-loaded system with two deployments.
So, he’s observing a noisy neighbor effect for frequency under TDP constraints.
If some cores use more power with AVX-512, the other cores may also see a frequency degradation.
I think even that may not be true starting with SPR, as the cores have autonomous power management.
But I am not 100% certain. I will ask and let you know.

Thanks!
Adrian 
"
2 for your concern about out range memory issue about "_mm_lddqu_si128":
1) first current stream mechinism can additional bytes for our processing . our decode process varintparse is private function and server for up message level  depend on the buffer from streaming processing. In sepecifct words the buffer is from ParseContext::Done processing(every object decode start it will call this func). Maybe you coulddeepdive in this process you will found some wondeful mechnism to prevent memory run out in some range .In this buffer size check and buffer swap processing ,there has some woderful mechinisim to  processing some visit out of range issue --at most additional 16 bytes (kSlopBytes)for run-out processing(in EpsCopyInputStream class that is base of ParseContext) . It afford additional  in "char buffer_[2 * kSlopBytes] = {};" . The specific processing is comlicated and hard to description very shortly.  So its  better to check that ParseContext::Done code line by line.
2) second read extra bytes didn't impact the algo correctless:
beacause the highest bytes 's msb always zero. even we get the additional undefined bytes ,this will not implact our correctless.

3 for your concern about pext execution effiency issue in AMD platform:
 May we need check if the new AMD platform has same issue. Or we could do some tricky processing for